+++ src/components/modals/ConfirmationModal.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
-import Modal from './modal';
+import Modal from './Modal';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import OutlinedButton from '../buttons/OutlinedButton';
 import Button from '../buttons/Button';
 import { browserPreferredTheme } from '../../theme';
+import Card from '../Card'
+import ModalOverlay from '../modals/ModalOverlay';
 
 const propTypes = {
   title: PropTypes.string.isRequired,
@@ -35,14 +37,43 @@ const ConfirmButton = styled(Button)`
   background-color: ${props => props.color}
 `;
 
+const StyledCard = styled(Card)`
+  display: grid;
+  grid-template-rows: [title] 1fr [content] 1.5fr;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%,-50%);
+  width: 100%;
+  max-width: 400px;
+  background-color: ${props => props.theme.black}
+`;
+
+const Title = styled.h2`
+  grid-row: title;
+  width: 75%;
+  margin: 0 auto;
+  text-align: center;
+`;
+
+const ContentWrapper = styled.div`
+  grid-row: content;
+  padding-top: 20px;
+`;
+
 export default function ConfirmationModal(props) {
   return(
-    <Modal visible={true} title={props.title}>
-      <Content>
-        <CancelButton onClick={props.cancelCallback}>cancel</CancelButton>
-        <ConfirmButton color={props.confirmationButtonColor} onClick={props.confirmationCallback}>{props.confirmationButtonText}</ConfirmButton>
-      </Content>
-    </Modal>
+    <ModalOverlay title={props.title}>
+      <StyledCard>
+        <Title>{props.title}</Title>
+        <ContentWrapper>
+          <Content>
+            <CancelButton onClick={props.cancelCallback}>cancel</CancelButton>
+            <ConfirmButton color={props.confirmationButtonColor} onClick={props.confirmationCallback}>{props.confirmationButtonText}</ConfirmButton>
+          </Content>
+        </ContentWrapper>
+      </StyledCard>
+    </ModalOverlay>
   );
 }
